+++ src/main/java/duke/task/TaskException.java
@@ -1,3 +1,4 @@
+package duke.task;
 
 public class TaskException extends Exception {
     public TaskException(String message) {
