+++ src/main/java/duke/util/Parser.java
@@ -1,3 +1,7 @@
+package duke.util;
+
+import duke.task.*;
+
 import java.util.Arrays;
 
 public class Parser {
