+++ src/main/java/duke/util/Ui.java
@@ -1,3 +1,8 @@
+package duke.util;
+
+import duke.task.Task;
+import duke.task.TaskList;
+
 import java.util.ArrayList;
 
 public class Ui {
