+++ text-ui-test/EXPECTED.TXT
@@ -4,19 +4,19 @@
     ____________________________________________________________
 
     ____________________________________________________________
-     Got it. I've added this Task:
+     Got it. I've added this duke.task.Task:
        [T][ ] borrow book
      Now you have 1 tasks in the list.
     ____________________________________________________________
 
     ____________________________________________________________
-     Got it. I've added this Task:
+     Got it. I've added this duke.task.Task:
        [D][ ] return book(by: Sunday)
      Now you have 2 tasks in the list.
     ____________________________________________________________
 
     ____________________________________________________________
-     Got it. I've added this Task:
+     Got it. I've added this duke.task.Task:
        [E][ ] project meeting (from: Mon 2pm to: 4pm)
      Now you have 3 tasks in the list.
     ____________________________________________________________
