+++ text-ui-test/input.txt
@@ -1,3 +1,3 @@
-Todo borrow book
-Deadline return book /by Sunday
-Event project meeting /from Mon 2pm /to 4pm
+duke.task.Todo borrow book
+duke.task.Deadline return book /by Sunday
+duke.task.Event project meeting /from Mon 2pm /to 4pm
